Add `FirestoreQueryTest.kt`
